@@ -1,72 +1,47 @@
 {
   "name": "spring-petclinic-reactjs",
-  "version": "0.0.1",
-  "description": "Spring Boot Pet Clinic Sample Application with React Frontend and Elastic. Forked from https://github.com/spring-petclinic/spring-petclinic-reactjs",
+  "version": "0.0.2",
+  "description": "Spring Petclinic React frontend (migrated to React 19 + Vite)",
   "main": "index.js",
   "repository": "https://github.com/elastic/spring-petclinic",
-  "scripts": {
-    "postinstall": "typings install",
-    "install-types": "typings install",
-    "test": "jest",
-    "test:watch": "jest --watchAll --no-cache",
-    "start": "webpack-dev-server --open",
-    "build:clean": "rm -rf ./dist && webpack --config webpack.config.js",
-    "build:prod": "rm -rf ./dist && NODE_ENV=production webpack --config webpack.config.prod.js"
-  },
   "author": "Nils Hartmann",
   "license": "ISC",
+  "private": true,
+  "engines": {
+    "node": ">=20.19"
+  },
+
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "typecheck": "tsc --noEmit",
+    "test": "vitest",
+    "test:run": "vitest run",
+    "test:ui": "vitest --ui"
+  },
+
   "dependencies": {
     "@elastic/apm-rum": "^5.1.1",
-    "bootstrap": "^3.3.7",
-    "classnames": "^2.2.6",
-    "moment": "^2.22.2",
-    "react": "^15.5.8",
-    "react-autosuggest": "^9.4.2",
-    "react-datepicker": "^0.29.0",
-    "react-dom": "^15.5.8",
-    "react-router": "^2.7.0",
-    "redbox-react": "^1.2.3",
-    "whatwg-fetch": "^1.0.0"
+    "bootstrap": "^3.4.1",
+    "classnames": "^2.5.1",
+    "moment": "^2.30.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.10.1",
+    "react-autosuggest": "^10.1.0",
+    "react-datepicker": "^4.24.0",
+    "whatwg-fetch": "^3.6.20"
   },
+
   "devDependencies": {
-    "ajv": "^6.5.4",
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "chalk": "^1.1.3",
-    "cross-env": "^1.0.7",
-    "css-loader": "^0.23.1",
-    "enzyme": "^2.5.1",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^16.0.1",
-    "jest-fetch-mock": "^1.0.6",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "react-addons-test-utils": "^15.3.2",
-    "react-hot-loader": "^3.0.0-beta.0",
-    "rimraf": "^2.5.0",
-    "style-loader": "^0.13.0",
-    "ts-jest": "^0.1.8",
-    "ts-loader": "^0.8.2",
-    "tslint": "^3.15.1",
-    "tslint-loader": "^2.1.5",
-    "typescript": "~2.0.2",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.2.0",
-    "webpack-dev-server": "^2.9.7",
-    "typings": "^1.3.2"
-  },
-  "jest": {
-    "scriptPreprocessor": "./node_modules/ts-jest/preprocessor.js",
-    "testRegex": "(/__tests__/.*\\.(test|spec))\\.(ts|tsx|js)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ]
+    "vite": "^7.3.0",
+    "@vitejs/plugin-react-swc": "^4.2.2",
+    "typescript": "^5.9.3",
+    "less": "^4.2.0",
+    "vitest": "^4.0.16",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/jest-dom": "^6.4.2",
+    "jsdom": "^24.0.0"
   }
 }
