@@ -1,15 +1,9 @@
-import * as React from 'react';
-const { Router } = require('react-router');
+import React from 'react';
+import { RouterProvider } from 'react-router-dom';
+import router from './configureRoutes';
 
-// Routes
-import configureRoutes from './configureRoutes';
-
-// https://github.com/gaearon/react-hot-boilerplate/pull/61#issuecomment-218333616
-// https://github.com/rybon/counter-hmr/blob/e651ce25b3a307f13ca53c977f9e8709ba873407/src/components/Root.jsx
-const Root = ({history}) => (
-    <Router history={history}>
-      {configureRoutes()}
-  </Router>
-);
+const Root: React.FC = () => {
+  return <RouterProvider router={router} />;
+};
 
 export default Root;
