@@ -1,101 +1,105 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
+// main.tsx
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+// React Router agora é fornecido pelo Root via RouterProvider
+import Root from './Root';
 
-import { browserHistory as history } from 'react-router';
+// Estilos (LESS). Vite cuida do pipeline via css.preprocessorOptions.
+import './styles/less/petclinic.less';
 
-const initApm = require('@elastic/apm-rum/src/index').init;
-
-require('./styles/less/petclinic.less');
+// APM RUM (ESM): mantemos mesma lógica de inicialização
+import { init as initApm } from '@elastic/apm-rum';
 import { url } from './util/index';
 
+// Exportando a classe APMService
 export class APMService {
-
   private static instance: APMService;
   private apm: any;
   private current_span: any;
   private span_open = false;
   private ready = false;
   private open = false;
-  private constructor () {
 
-  }
+  private constructor() {}
 
-  private setup_apm() {
-      const requestUrl = url('config');
-      const xhr = new XMLHttpRequest();
-      xhr.open('GET', requestUrl, true);
-      xhr.onload = function(e) {
-        if (xhr.status === 200) {
-            const config = JSON.parse(xhr.responseText);
-            this.apm = initApm({
-               serviceName: config.apm_client_service_name,
-               serverUrl: config.apm_server_js,
-               serviceVersion: config.apm_service_version,
-               transactionThrottleLimit: 1000,
-               errorThrottleLimit: 1000,
-               distributedTracingOrigins: config.distributedTracingOrigins.split(',')
+  private async setup_apm() {
+    const requestUrl = url('config');
+    try {
+      const resp = await fetch(requestUrl);
+      if (!resp.ok) {
+        console.log('Failed to Initialize APM');
+        console.log(`Failed GET on ${requestUrl} - ${resp.status} ${resp.statusText}`);
+        return;
+      }
+      const config = await resp.json();
+      this.apm = initApm({
+        serviceName: config.apm_client_service_name,
+        serverUrl: config.apm_server_js,
+        serviceVersion: config.apm_service_version,
+        transactionThrottleLimit: 1000,
+        errorThrottleLimit: 1000,
+		logLevel: 'debug', // >>> DEBUG
+        distributedTracingOrigins: (config.distributedTracingOrigins || '').split(',')
+      });
+      this.apm.setInitialPageLoadName(
+        window.location.pathname !== '' ? window.location.pathname : 'homepage'
+      );
+      this.apm.addFilter(function (payload: any) {
+        if (payload.transactions) {
+          payload.transactions.filter(function (tr: any) {
+            return tr.spans.some(function (span: any) {
+              return (
+                span.context &&
+                span.context.http &&
+                span.context.http.url &&
+                (span.context.http.url.includes('rum/transactions') ||
+                 span.context.http.url.includes('rum/events'))
+              );
             });
-            this.apm.setInitialPageLoadName(window.location.pathname !== '' ? window.location.pathname : 'homepage');
-            this.apm.addFilter(function (payload) {
-              if (payload.transactions) {
-                payload.transactions.filter(function (tr) {
-                  return tr.spans.some(function (span) {
-                    return (span.context && span.context.http && span.context.http.url && (span.context.http.url.includes('rum/transactions')
-                    || span.context.http.url.includes('rum/events')));
-                  });
-                });
-              };
-              return payload;
-            });
-            this.apm.setUserContext({
-                'username': config.user.username,
-                'email': config.user.email
-            });
-            APMService.instance.ready = true;
-        } else {
-          console.log('Failed to Initialize APM');
-          console.log(`Failed GET on ${requestUrl} - ${xhr.status} ${xhr.statusText}`);
+          });
         }
-      }.bind(this);
-      xhr.onerror = function(e) {
-         console.log('Failed to Initialize APM');
-         console.log(`Failed GET on ${requestUrl} - ${xhr.status} ${xhr.statusText}`);
-      };
-      xhr.send(null);
+        return payload;
+      });
+      this.apm.setUserContext({
+        username: config.user?.username,
+        email: config.user?.email
+      });
+      this.ready = true;
+    } catch (e) {
+      console.log('Failed to Initialize APM');
+      console.log(e);
+    }
   }
 
   static getInstance() {
     if (!APMService.instance) {
-        console.log('Creating APM Service');
-        APMService.instance = new APMService();
-        APMService.instance.setup_apm();
-        console.log('Created APM Service');
+      console.log('Creating APM Service');
+      APMService.instance = new APMService();
+      APMService.instance.setup_apm();
+      console.log('Created APM Service');
     }
     return APMService.instance;
   }
 
-
-  startTransaction(name) {
-    if (APMService.instance.ready && !APMService.instance.open) {
+  startTransaction(name: string) {
+    if (this.ready && !this.open) {
       console.log('Starting transaction - ' + name + ':');
-      // in case one has been opened
-      if (APMService.instance.apm.getCurrentTransaction()) {
-        APMService.instance.apm.getCurrentTransaction().end();
+      if (this.apm.getCurrentTransaction()) {
+        this.apm.getCurrentTransaction().end();
       }
-      let transaction = APMService.instance.apm.startTransaction(name, 'Events');
-      APMService.instance.apm.addLabels('success_load', false);
+      const transaction = this.apm.startTransaction(name, 'Events');
+      this.apm.addLabels('success_load', false);
       console.log(transaction);
-      APMService.instance.open = true;
+      this.open = true;
     }
   }
 
-  endTransaction(completed) {
-    if (APMService.instance.open) {
-      APMService.instance.open = false;
-      APMService.instance.apm.addLabels('success_load', completed.toString());
+  endTransaction(completed: boolean) {
+    if (this.open) {
+      this.open = false;
+      this.apm.addLabels('success_load', completed.toString());
       console.log('Closing transaction');
-      let transaction = APMService.instance.apm.getCurrentTransaction();
+      const transaction = this.apm.getCurrentTransaction();
       if (transaction) {
         transaction.end();
         console.log('Closed transaction:');
@@ -104,96 +108,97 @@ export class APMService {
     }
   }
 
-  startSpan(name, type) {
-    if (APMService.instance.ready && APMService.instance.open) {
-      let transaction = APMService.instance.apm.getCurrentTransaction();
-      APMService.instance.span_open = true;
-      APMService.instance.current_span = transaction.startSpan(name, type);
+  startSpan(name: string, type: string) {
+    if (this.ready && this.open) {
+      const transaction = this.apm.getCurrentTransaction();
+      this.span_open = true;
+      this.current_span = transaction.startSpan(name, type);
     }
   }
 
   endSpan() {
-    if (APMService.instance.open && APMService.instance.span_open) {
-      APMService.instance.current_span.end();
-      APMService.instance.span_open = false;
+    if (this.open && this.span_open) {
+      this.current_span.end();
+      this.span_open = false;
     }
   }
 
-  captureError(message) {
-    if (APMService.instance.open) {
+  captureError(message: string) {
+    if (this.open) {
       console.log('Capturing Error');
-      APMService.instance.apm.captureError(new Error(message));
+      this.apm.captureError(new Error(message));
     }
   }
-
 }
 
+// (Opcional) Inicializar APM aqui, se quiser estar ativo desde o boot:
+APMService.getInstance();
+
+
+// >>> DEBUG: rastrear chamadas fetch no browser
+if ((window as any).__DEBUG_HTTP || import.meta.env.MODE !== 'production') {
+  const origFetch = window.fetch.bind(window);
+  window.fetch = async (input: RequestInfo | URL, init?: RequestInit) => {
+    const url = typeof input === 'string' ? input : (input as any).url;
+    const method = (init?.method || 'GET').toUpperCase();
+    const t0 = performance.now();
+    try {
+      const resp = await origFetch(input, init);
+      const dt = (performance.now() - t0).toFixed(1);
+      console.log(`[BROWSER] ${method} ${url} -> ${resp.status} (${dt}ms)`);
+      return resp;
+    } catch (e) {
+      const dt = (performance.now() - t0).toFixed(1);
+      console.error(`[BROWSER] ${method} ${url} -> ERROR (${dt}ms)`, e);
+      throw e;
+    }
+  };
+}
+// <<< DEBUG
+
+
+// Render da aplicação (React 19 + createRoot)
+const mountPoint = document.getElementById('mount')!;
+createRoot(mountPoint).render(
+  <React.StrictMode>
+    <Root />
+  </React.StrictMode>
+);
+
+// Legado de IE: mantemos a função se algum componente a usar
 export const detectIE = () => {
-    let ua = window.navigator.userAgent;
-    let msie = ua.indexOf('MSIE ');
-    if (msie > 0) {
-        // IE 10 or older => return version number
-        return parseInt(ua.substring(msie + 5, ua.indexOf('.', msie)), 10);
-    }
-    let trident = ua.indexOf('Trident/');
-    if (trident > 0) {
-        // IE 11 => return version number
-        let rv = ua.indexOf('rv:');
-        return parseInt(ua.substring(rv + 3, ua.indexOf('.', rv)), 10);
-    }
-    let edge = ua.indexOf('Edge/');
-    if (edge > 0) {
-       // Edge (IE 12+) => return version number
-       return parseInt(ua.substring(edge + 5, ua.indexOf('.', edge)), 10);
-    }
-    // other browser
-    return -1;
+  const ua = window.navigator.userAgent;
+  const msie = ua.indexOf('MSIE ');
+  if (msie > 0) return parseInt(ua.substring(msie + 5, ua.indexOf('.', msie)), 10);
+  const trident = ua.indexOf('Trident/');
+  if (trident > 0) {
+    const rv = ua.indexOf('rv:');
+    return parseInt(ua.substring(rv + 3, ua.indexOf('.', rv)), 10);
+  }
+  const edge = ua.indexOf('Edge/');
+  if (edge > 0) return parseInt(ua.substring(edge + 5, ua.indexOf('.', edge)), 10);
+  return -1;
 };
 
 export const pi = (count: number) => {
-    let inside = 0;
-    for (let i = 0; i < count; i++) {
-        let x = (Math.random() * 2) - 1;
-        let y = (Math.random() * 2) - 1;
-        if (((x * x) + (y * y)) < 1) {
-            inside++;
-        }
-    }
-    return 4.0 * (inside / count);
+  let inside = 0;
+  for (let i = 0; i < count; i++) {
+    const x = Math.random() * 2 - 1;
+    const y = Math.random() * 2 - 1;
+    if (x * x + y * y < 1) inside++;
+  }
+  return 4.0 * (inside / count);
 };
 
 export const punish = () => {
-    if (detectIE() > -1) {
-        console.log('Anyone who uses IE deserves to be punished!');
-        let pain = 50000000 + Math.floor(Math.random() * 25000000);
-        console.log('Amount of Pain: ' + pain);
-        let val = pi(pain);
-        console.log(val);
-    };
+  if (detectIE() > -1) {
+    console.log('Anyone who uses IE deserves to be punished!');
+    const pain = 50_000_000 + Math.floor(Math.random() * 25_000_000);
+    console.log('Amount of Pain: ' + pain);
+    const val = pi(pain);
+    console.log(val);
+  }
 };
 
-/**
-
- **/
-// The Application
-import Root from './Root';
-
-// Render Application
-const mountPoint = document.getElementById('mount');
-ReactDOM.render(
-  <AppContainer><Root history={history}/></AppContainer>,
-  mountPoint
-);
-
-declare var module: any;
-if (module.hot) {
-  module.hot.accept('./Root', () => {
-    const NextApp = require('./Root').default;
-    ReactDOM.render(
-      <AppContainer>
-        <NextApp history={history} />
-      </AppContainer>,
-      mountPoint
-    );
-  });
-}
+// Exportando também a instância singleton para fácil acesso
+export const apmServiceInstance = APMService.getInstance();
