@@ -1,145 +1,176 @@
-const { Chromeless } = require('chromeless')
+// tasks.js
 
-var args = process.argv.slice(2);
-var users_file = args[0];
-var delay_per_user = args[1];
-var delay_per_page = args[2];
-delay_per_page = delay_per_page.split(':')
-delay_per_user = delay_per_user.split(':')
-//read variables
+const puppeteer = require('puppeteer');
+const csv = require('csvtojson');
+
+// Lê argumentos da linha de comando
+const args = process.argv.slice(2);
+const usersFile = args[1]; // Corrigido: primeiro argumento é o arquivo de usuários
+const delayPerUser = args[2]; // ex: "20000:10000"
+const delayPerPage = args[3]; // ex: "120000:20000"
+
+if (!usersFile || !delayPerUser || !delayPerPage) {
+  console.error('Uso: node tasks.js <users.csv> <delay_per_user> <delay_per_page>');
+  console.error('Ex: node tasks.js users.csv 20000:10000 120000:20000');
+  process.exit(1);
+}
+
+const [delayPerUserAvg, delayPerUserDev] = delayPerUser.split(':').map(Number);
+const [delayPerPageAvg, delayPerPageDev] = delayPerPage.split(':').map(Number);
+
+// Validação
+if (isNaN(delayPerUserAvg) || isNaN(delayPerPageAvg)) {
+  console.error('Delays inválidos. Use formato: N:M (ex: 20000:10000)');
+  console.error(`Users file: ${usersFile}, Delay Per User: ${delayPerUser}, Delay Per Page: ${delayPerPage}`);
+  process.exit(1);
+}
+
+const baseUrl = process.env.PETCLINIC_BASE_URL || 'http://spring-petclinic-client:3000';
 console.log('Starting loadgenerator...');
-console.log('Users file: '+ users_file + ', Delay Per User: '+ delay_per_user + ', Delay Per Page: ' + delay_per_page)
+console.log(`Users file: ${usersFile}, Delay Per User: ${delayPerUser}, Delay Per Page: ${delayPerPage}`);
+console.log(JSON.stringify(baseUrl));
 
-
-const csv=require('csvtojson')
-var baseUrl = process.env.PETCLINIC_BASE_URL || 'http://localhost:8081';
-var url = baseUrl
-const chromeless = new Chromeless({ launchChrome: true,  waitTimeout: 30000, scrollBeforeClick: true, implicitWait: true });
-
+// Função para ler CSV
 async function readCSV(filename) {
   return await csv({
-    noheader:true,
-    headers: ['username','ip','agent']
+    noheader: true,
+    headers: ['username', 'ip', 'agent']
   }).fromFile(filename);
 }
 
-function sleep(ms){
-    return new Promise(resolve=>{
-        setTimeout(resolve,ms)
-    })
+// Função sleep
+function sleep(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
 }
 
-async function run(username, ip, agent) {
-    url = await chromeless
-    .setUserAgent(agent)
-    .goto(url)
-    .setExtraHTTPHeaders({
-        'x-forwarded-for': ip,
-        'x-forwarded-user': username,
-        'x-forwarded-ip': ip,
-        'user-agent': agent
-    })
-    .evaluate((url,baseUrl,delay_per_page_avg, delay_per_page_dev) => {
-        var delay  = delay_per_page_avg + (Math.random() * delay_per_page_dev);
-        console.log(delay);
-        var timer = window.setInterval(function () {
-            //navigates around page dynamically
-            var links = document.querySelectorAll('a[href^="/"]');
-            var uniq_links = {};
-            var max_length = 0;
-            var min_length = Number.MAX_SAFE_INTEGER;
-            // we attempt to bias longer urls
-            for ( var i=0, len=links.length; i < len; i++ ) {
-                uniq_links[links[i].href] = links[i];
-                if (links[i].href.length < min_length) {
-                    min_length = links[i].href.length;
-                } else if (links[i].href.length > max_length) {
-                    max_length = links[i].href.length;
-                }
-            }
-            var diff = 10/(max_length - min_length);
-            links = new Array();
-            //refs = new Array();
-            for ( var key in uniq_links ) {
-                for (i = 0; i < (uniq_links[key].href.length - min_length + 1) * diff ; i++){
-                    links.push(uniq_links[key]);
-                    //refs.push(uniq_links[key].href);
-                }
-            }
-            //console.log(refs);
-            if (links && links.length) {
-                var i = Math.floor(Math.random()*links.length);
-                //console.log(links[i])
-                links[i].click();
-            } else {
-                //no where to go, wait for page reload
-                window.clearInterval(timer);
-            }
-        }, delay );
-        var uniq_links = {};
-        var links = document.querySelectorAll('a[href^="/"]');
-        for ( var i=0, len=links.length; i < len; i++ ) {
-            uniq_links[links[i].href] = links[i];
+// Função principal de navegação
+async function run(browser, username, ip, agent, currentUrl) {
+  const page = await browser.newPage();
+
+  // Configura headers e user agent
+  await page.setExtraHTTPHeaders({
+    'x-forwarded-for': ip,
+    'x-forwarded-user': username,
+    'x-forwarded-ip': ip
+  });
+  await page.setUserAgent(agent);
+
+  try {
+    await page.goto(currentUrl, { waitUntil: 'networkidle2', timeout: 30000 });
+
+    // Avalia no navegador: navegação aleatória com bias para URLs mais longas
+    const newUrl = await page.evaluate((baseUrl, avg, dev) => {
+      const delay = avg + Math.floor(Math.random() * dev);
+      let timer;
+
+      const navigate = () => {
+        const links = Array.from(document.querySelectorAll('a[href^="/"]'));
+        const uniq = new Map();
+        let minLen = Infinity;
+        let maxLen = 0;
+
+        for (const link of links) {
+          if (!link.href) continue;
+          uniq.set(link.href, link);
+          const len = link.href.length;
+          if (len < minLen) minLen = len;
+          if (len > maxLen) maxLen = len;
         }
-        links = new Array();
-        for ( var key in uniq_links ) {
-            links.push(uniq_links[key]);
+
+        if (uniq.size === 0) {
+          clearInterval(timer);
+          return baseUrl;
         }
-        if (links && links.length) {
-            return links[Math.floor(Math.random()*links.length)].href;
-        } else {
-            return baseUrl;
+
+        // Bias para links mais longos
+        const diff = maxLen === minLen ? 1 : 10 / (maxLen - minLen);
+        const weightedLinks = [];
+        for (const [href, link] of uniq.entries()) {
+          const weight = Math.max(1, Math.floor((href.length - minLen + 1) * diff));
+          for (let i = 0; i < weight; i++) {
+            weightedLinks.push(link);
+          }
         }
-    }, url, baseUrl, parseInt(delay_per_page[0]), parseInt(delay_per_page[1]));
-    console.log(url);
+
+        if (weightedLinks.length > 0) {
+          const randomLink = weightedLinks[Math.floor(Math.random() * weightedLinks.length)];
+          randomLink.click();
+          return randomLink.href;
+        }
+
+        return baseUrl;
+      };
+
+      // Navega uma vez imediatamente
+      const result = navigate();
+
+      // Agenda navegações repetidas
+      timer = setInterval(navigate, delay);
+      // Obs: o timer continua rodando, mas não precisamos retornar URL dinâmica
+
+      return result;
+    }, baseUrl, delayPerPageAvg, delayPerPageDev);
+
+    // Aguarda um tempo antes de fechar (simula usuário navegando)
+    await sleep(delayPerPageAvg + Math.floor(Math.random() * delayPerPageDev));
+
+    // Fecha a página para liberar memória
+    await page.close();
+    return newUrl || baseUrl;
+  } catch (error) {
+    console.error(`Erro ao processar URL ${currentUrl}:`, error.message);
+    await page.close();
+    return baseUrl;
+  }
 }
 
-async function safe_run() {
-      console.log('Starting from baseurl: '+url)
-      const users = await readCSV(users_file);
-      console.log(`Loaded ${users.length} users`)
-      for(;;) {
-          var user = users[Math.floor(Math.random() * users.length)]
-          console.log('Handing ' + url)
-          run(user.username,user.ip,user.agent).catch(async (error) => {
-              console.log("failed for url")
-              console.log(error);
-          });
-          await sleep(parseInt(delay_per_user[0]) + Math.floor(Math.random()*parseInt(delay_per_user[1])));
-      }
-}
+// Loop principal
+async function safeRun() {
+  console.log(`Starting from baseurl: ${baseUrl}`);
+  const users = await readCSV(usersFile);
+  console.log(`Loaded ${users.length} users`);
 
-function selectRandomIP() {
-    return RANDOM_IPS[Math.floor(Math.random() * RANDOM_IPS.length)];
-}
+  // Inicia o navegador **uma única vez** (eficiência)
+  const browser = await puppeteer.launch({
+    executablePath: process.env.CHROME_PATH || '/usr/bin/google-chrome',
+    headless: true,
+    args: [
+      '--no-sandbox',
+      '--disable-setuid-sandbox',
+      '--disable-dev-shm-usage',
+      '--disable-gpu',
+      '--single-process'
+    ]
+  });
 
-function loadRandomIPs() {
-    var IPs = [];
-    while (IPs.length < NUM_IPS) {
-      var randomIP = randomIp();
-      if (IPs.indexOf(randomIP) === -1) {
-        IPs.push(randomIP);
-      }
+  let currentUrl = baseUrl;
+
+  for (;;) {
+    const user = users[Math.floor(Math.random() * users.length)];
+    console.log(`Handling ${currentUrl} for user ${user.username}`);
+
+    try {
+      currentUrl = await run(browser, user.username, user.ip, user.agent, currentUrl);
+    } catch (error) {
+      console.error('Falha na execução do usuário:', error);
+      currentUrl = baseUrl; // Reinicia
     }
-    return IPs;
+
+    // Delay entre usuários
+    const userDelay = delayPerUserAvg + Math.floor(Math.random() * delayPerUserDev);
+    await sleep(userDelay);
+  }
 }
 
+// Tratamento de erros globais
+safeRun().catch(err => {
+  console.error('Erro fatal no load generator:', err);
+  process.exit(1);
+});
 
-function randomByte () {
-  return Math.round(Math.random()*256);
-}
-
-function isPrivate(ip) {
-  return /^10\.|^192\.168\.|^172\.16\.|^172\.17\.|^172\.18\.|^172\.19\.|^172\.20\.|^172\.21\.|^172\.22\.|^172\.23\.|^172\.24\.|^172\.25\.|^172\.26\.|^172\.27\.|^172\.28\.|^172\.29\.|^172\.30\.|^172\.31\./.test(ip);
-}
-
-function randomIp() {
-  var ip = randomByte() +'.' +
-           randomByte() +'.' +
-           randomByte() +'.' +
-           randomByte();
-  if (isPrivate(ip)) { return randomIp(); }
-  return ip;
-}
-
-safe_run().catch(console.error.bind(console))
+// Encerra o navegador se o processo for interrompido
+process.on('SIGINT', async () => {
+  console.log('Encerrando...');
+  await browser?.close();
+  process.exit(0);
+});
