@@ -1,21 +1,31 @@
-import * as React from 'react';
+
+import React from 'react';
+import { Outlet, useLocation } from 'react-router-dom';
 import Menu from './Menu';
 
-export default ({location, children}) => (
-  <div>
-    <Menu name={location.pathname} />
-    <div className='container-fluid'>
-      <div className='container xd-container'>
+const App: React.FC = () => {
+  const location = useLocation();
 
-        {children}
+  return (
+    <div>
+      <Menu name={location.pathname} />
+      <div className="container-fluid">
+        <div className="container xd-container">
 
-        <div className='container'>
-          <div className='row'>
-            <div className='col-12 text-center'>
-              <img src='/images/spring-pivotal-logo.png' alt='Sponsored by Pivotal' /></div>
+          <Outlet />
+
+          <div className="container">
+            <div className="row">
+              <div className="col-12 text-center">
+                <img src="/images/spring-pivotal-logo.png" alt="Sponsored by Pivotal" />
+              </div>
+            </div>
           </div>
+
         </div>
       </div>
     </div>
-  </div>
-);
+  );
+};
+
+export default App;
