@@ -1,48 +1,52 @@
-import * as React from 'react';
-import { Link } from 'react-router';
 
+import React from 'react';
+import { NavLink } from 'react-router-dom';
 
-const MenuItem = ({active, url, title, children}: { active: boolean, url: string, title: string, children?: any }) => (
-  <li className={active ? 'active' : ''}>
-    <Link to={url}>{children}</Link>
+const MenuItem: React.FC<{ url: string; title: string; children?: React.ReactNode }> = ({ url, title, children }) => (
+  <li>
+    <NavLink
+      to={url}
+      className={({ isActive }) => (isActive ? 'active' : undefined)}
+      aria-label={title}
+    >
+      {children}
+    </NavLink>
   </li>
 );
 
-export default ({name}: { name: string }) => (
-  <nav className='navbar navbar-default' role='navigation'>
-    <div className='container'>
-      <div className='navbar-header'>
-        <a className='navbar-brand' href='/'><span></span></a>
-        <button type='button' className='navbar-toggle' data-toggle='collapse' data-target='#main-navbar'>
-          <span className='icon-bar'></span>
-          <span className='icon-bar'></span>
-          <span className='icon-bar'></span>
+const Menu: React.FC<{ name?: string }> = () => (
+  <nav className="navbar navbar-default" role="navigation">
+    <div className="container">
+      <div className="navbar-header">
+        <a className="navbar-brand" href="/"><span></span></a>
+        <button type="button" className="navbar-toggle" data-toggle="collapse" data-target="#main-navbar">
+          <span className="icon-bar"></span>
+          <span className="icon-bar"></span>
+          <span className="icon-bar"></span>
         </button>
       </div>
-      <div className='navbar-collapse collapse' id='main-navbar'>
-        <ul className='nav navbar-nav navbar-right'>
-          <MenuItem active={name === '/'} url='/' title='home page'>
-            <span className='glyphicon glyphicon-home' aria-hidden='true'></span>&nbsp;
-                    <span>Home</span>
+
+      <div className="navbar-collapse collapse" id="main-navbar">
+        <ul className="nav navbar-nav navbar-right">
+          <MenuItem url="/" title="home page">
+            <span className="glyphicon glyphicon-home" aria-hidden="true"></span>&nbsp;<span>Home</span>
           </MenuItem>
 
-          <MenuItem active={name.startsWith('/owners')} url='/owners/list' title='find owners'>
-            <span className='glyphicon glyphicon-search' aria-hidden='true'></span>&nbsp;
-                    <span>Find owners</span>
+          <MenuItem url="/owners/list" title="find owners">
+            <span className="glyphicon glyphicon-search" aria-hidden="true"></span>&nbsp;<span>Find owners</span>
           </MenuItem>
 
-          <MenuItem active={name === 'vets'} url='/vets' title='veterinarians'>
-            <span className='glyphicon glyphicon-th-list' aria-hidden='true'></span>&nbsp;
-                    <span>Veterinarians</span>
+          <MenuItem url="/vets" title="veterinarians">
+            <span className="glyphicon glyphicon-th-list" aria-hidden="true"></span>&nbsp;<span>Veterinarians</span>
           </MenuItem>
 
-          <MenuItem active={name === 'error'} url='/error'
-            title='trigger a RuntimeException to see how it is handled'>
-            <span className='glyphicon glyphicon-warning-sign' aria-hidden='true'></span>&nbsp;
-                    <span>Error</span>
+          <MenuItem url="/error" title="trigger a RuntimeException to see how it is handled">
+            <span className="glyphicon glyphicon-warning-sign" aria-hidden="true"></span>&nbsp;<span>Error</span>
           </MenuItem>
         </ul>
       </div>
     </div>
   </nav>
 );
+
+export default Menu;
