@@ -1,3 +1,5 @@
-import * as React from 'react';
+
+import React from 'react';
 
 export default () => <h3>The requested page has not been found.</h3>;
+
