@@ -1,140 +1,106 @@
-import * as React from 'react';
-import { IRouter, Link } from 'react-router';
-import { IOwner, IRouterContext } from '../../types/index';
-import { request, xhr_request } from '../../util/index';
+
+import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
+
+import { IOwner } from '../../types/index';
+import { xhr_request } from '../../util/index';
 import { APMService, punish } from '../../main';
 import OwnersTable from './OwnersTable';
 
-interface IFindOwnersPageProps {
-  location: HistoryModule.Location;
-}
-
-interface IFindOwnersPageState {
-  owners?: IOwner[];
-  filter?: string;
-}
-
-const getFilterFromLocation = (location) => {
-  return location.query ? (location.query as any).lastName : null;
+const useFilterFromLocation = (location: ReturnType<typeof useLocation>) => {
+  const params = new URLSearchParams(location.search);
+  return params.get('lastName');
 };
 
-export default class FindOwnersPage extends React.Component<IFindOwnersPageProps, IFindOwnersPageState> {
-  context: IRouterContext;
-  initial_render: boolean;
+const FindOwnersPage: React.FC = () => {
+  const location = useLocation();
+  const navigate = useNavigate();
+  const initialRender = useRef(true);
 
-  static contextTypes = {
-    router: React.PropTypes.object.isRequired
-  };
+  const [owners, setOwners] = useState<IOwner[] | undefined>(undefined);
+  const [filter, setFilter] = useState<string | null>(useFilterFromLocation(location));
 
+  const fetchData = useCallback((f: string | null) => {
+    const query = f ? encodeURIComponent(f) : '';
+    const requestUrl = f && query !== '*' ? `api/owners/*/lastname/${query}` : 'api/owners';
 
-
-  constructor(props) {
-    super(props);
-    APMService.getInstance().startTransaction('FindOwnersPage');
-    punish();
-    this.initial_render = true;
-    this.onFilterChange = this.onFilterChange.bind(this);
-    this.submitSearchForm = this.submitSearchForm.bind(this);
-
-    this.state = {
-      filter: getFilterFromLocation(props.location)
-    };
-  }
-
-  componentDidUpdate() {
-    if (this.initial_render) {
-      APMService.getInstance().endSpan();
-      APMService.getInstance().endTransaction(true);
-    }
-    this.initial_render = false;
-  }
-
-  componentWillUnmount() {
-    APMService.getInstance().endSpan();
-    APMService.getInstance().endTransaction(false);
-  }
-
-  componentDidMount() {
-    const { filter } = this.state;
-    this.fetchData(filter);
-  }
-
-  componentWillReceiveProps(nextProps: IFindOwnersPageProps) {
-    const { location } = nextProps;
-
-    // read the filter from uri
-    const filter = getFilterFromLocation(location);
-
-    // set state
-    this.setState({ filter });
-
-    // load data according to filter
-    this.fetchData(filter);
-  }
-
-  onFilterChange(event) {
-    this.setState({
-      filter: event.target.value as string
-    });
-  }
-
-  /**
-   * Invoked when the submit button was pressed.
-   *
-   * This method updates the URL with the entered lastName. The change of the URL
-   * leads to new properties and thus results in rerending
-   */
-  submitSearchForm() {
-    const { filter } = this.state;
-    APMService.getInstance().startTransaction('FindOwnersPage: Filter');
-    this.context.router.push({
-      pathname: '/owners/list',
-      query: { 'lastName': filter || '' }
-    });
-  }
-
-  /**
-   * Actually loads data from the server
-   */
-  fetchData(filter: string) {
-    const query = encodeURIComponent(filter);
-    const requestUrl = filter && query !== '*' ? 'api/owners/*/lastname/' + query : 'api/owners';
-    xhr_request(requestUrl, (status, owners) =>  {
+    xhr_request(requestUrl, (status, data) => {
       if (status < 400) {
         APMService.getInstance().startSpan('Page Render', 'react');
-        this.setState({ owners });
+        setOwners(data as IOwner[]);
       }
     });
-  }
+  }, []);
 
-  render() {
-    const { filter, owners } = this.state;
+  useEffect(() => {
+    APMService.getInstance().startTransaction('FindOwnersPage');
+    punish();
+    fetchData(filter);
 
-    return (
-      <span>
-        <section>
-          <h2>Find Owners</h2>
+    return () => {
+      APMService.getInstance().endSpan();
+      APMService.getInstance().endTransaction(false);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
-          <form className='form-horizontal' action='javascript:void(0)'>
-            <div className='form-group'>
-              <div className='control-group' id='lastName'>
-                <label className='col-sm-2 control-label'>Last name </label>
-                <div className='col-sm-10'>
-                  <input className='form-control' name='filter' value={filter || ''} onChange={this.onFilterChange} size={30} maxLength={80} />
-                  { /* <span className='help-inline'><form:errors path='*'/></span> TODO */}
-                </div>
+  // Reagir à mudança de querystring (ex.: lastName) e buscar novamente
+  useEffect(() => {
+    const f = useFilterFromLocation(location);
+    setFilter(f);
+    fetchData(f);
+
+    if (initialRender.current) {
+      APMService.getInstance().endSpan();
+      APMService.getInstance().endTransaction(true);
+      initialRender.current = false;
+    }
+  }, [location.search, fetchData]);
+
+  const onFilterChange = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
+    setFilter(event.target.value);
+  }, []);
+
+  const submitSearchForm = useCallback(() => {
+    APMService.getInstance().startTransaction('FindOwnersPage: Filter');
+    navigate(`/owners/list?lastName=${encodeURIComponent(filter || '')}`);
+  }, [navigate, filter]);
+
+  return (
+    <span>
+      <section>
+        <h2>Find Owners</h2>
+
+        <form className="form-horizontal" onSubmit={(e) => { e.preventDefault(); submitSearchForm(); }}>
+          <div className="form-group">
+            <div className="control-group" id="lastName">
+              <label className="col-sm-2 control-label">Last name </label>
+              <div className="col-sm-10">
+                <input
+                  className="form-control"
+                  name="filter"
+                  value={filter || ''}
+                  onChange={onFilterChange}
+                  size={30}
+                  maxLength={80}
+                />
               </div>
             </div>
-            <div className='form-group'>
-              <div className='col-sm-offset-2 col-sm-10'>
-                <button type='button' onClick={this.submitSearchForm} className='btn btn-default'>Find Owner</button>
-              </div>
+          </div>
+          <div className="form-group">
+            <div className="col-sm-offset-2 col-sm-10">
+              <button type="button" onClick={submitSearchForm} className="btn btn-default">
+                Find Owner
+              </button>
             </div>
-          </form>
-        </section>
-        <OwnersTable owners={owners} />
-        <Link className='btn btn-default' to='/owners/new'> Add Owner</Link>
-      </span>
-    );
-  }
+          </div>
+        </form>
+      </section>
+
+      <OwnersTable owners={owners || []} />
+      <Link className="btn btn-default" to="/owners/new">Add Owner</Link>
+    </span>
+  );
 };
+
+export default FindOwnersPage;
