@@ -1,9 +1,10 @@
-import * as React from 'react';
-import OwnerEditor from './OwnerEditor';
 
+import React from 'react';
+import OwnerEditor from './OwnerEditor';
 import { IOwner } from '../../types/index';
 
 const newOwner = (): IOwner => ({
+  // @ts-expect-error: id nulo em criação (tipagem original usa number)
   id: null,
   isNew: true,
   firstName: '',
