@@ -1,13 +1,13 @@
-import * as React from 'react';
 
-import { Link } from 'react-router';
+import React from 'react';
+import { Link } from 'react-router-dom';
 import { IOwner } from '../../types/index';
 
-export default ({owner}: { owner: IOwner }) => (
+export default ({ owner }: { owner: IOwner }) => (
   <section>
     <h2>Owner Information</h2>
 
-    <table className='table table-striped'>
+    <table className="table table-striped">
       <tbody>
         <tr>
           <th>Name</th>
@@ -28,8 +28,8 @@ export default ({owner}: { owner: IOwner }) => (
       </tbody>
     </table>
 
-    <Link to={`/owners/${owner.id}/edit`} className='btn btn-default'>Edit Owner</Link>
+    <Link to={`/owners/${owner.id}/edit`} className="btn btn-default">Edit Owner</Link>
     &nbsp;
-    <Link to={`/owners/${owner.id}/pets/new`} className='btn btn-default'>Add New Pet</Link>
+    <Link to={`/owners/${owner.id}/pets/new`} className="btn btn-default">Add New Pet</Link>
   </section>
 );
