@@ -1,10 +1,10 @@
-import * as React from 'react';
 
-import { Link } from 'react-router';
+import React from 'react';
+import { Link } from 'react-router-dom';
 import { IOwner, IPet } from '../../types/index';
 
-const VisitsTable = ({ownerId, pet}: { ownerId: number, pet: IPet }) => (
-  <table className='table-condensed'>
+const VisitsTable = ({ ownerId, pet }: { ownerId: number; pet: IPet }) => (
+  <table className="table-condensed">
     <thead>
       <tr>
         <th>Visit Date</th>
@@ -30,15 +30,15 @@ const VisitsTable = ({ownerId, pet}: { ownerId: number, pet: IPet }) => (
   </table>
 );
 
-export default ({owner}: { owner: IOwner }) => (
+export default ({ owner }: { owner: IOwner }) => (
   <section>
     <h2>Pets and Visits</h2>
-    <table className='table table-striped'>
+    <table className="table table-striped">
       <tbody>
         {owner.pets.map(pet => (
           <tr key={pet.id}>
-            <td style={{ 'verticalAlign': 'top' }}>
-              <dl className='dl-horizontal'>
+            <td style={{ verticalAlign: 'top' }}>
+              <dl className="dl-horizontal">
                 <dt>Name</dt>
                 <dd>{pet.name}</dd>
                 <dt>Birth Date</dt>
@@ -47,7 +47,7 @@ export default ({owner}: { owner: IOwner }) => (
                 <dd>{pet.type.name}</dd>
               </dl>
             </td>
-            <td style={{ 'verticalAlign': 'top' }}>
+            <td style={{ verticalAlign: 'top' }}>
               <VisitsTable ownerId={owner.id} pet={pet} />
             </td>
           </tr>
