@@ -1,13 +1,17 @@
-import { IRouter } from 'react-router';
-
 // ------------------------------------ ROUTER ------------------------------------
+// Tipo mínimo para manter compatibilidade de build até migrarmos tudo para hooks do RRv7.
+export interface IRouter {
+  push: (path: string) => void;
+  replace?: (path: string) => void;
+  goBack?: () => void;
+}
+
 export interface IRouterContext {
   router: IRouter;
-};
+}
 
 // ------------------------------------ UTIL --------------------------------------
-export type IHttpMethod = 'POST' | 'PUT' | 'GET';
-
+export type IHttpMethod = 'POST' | 'PUT' | 'GET' | 'PATCH' | 'DELETE';
 
 // ------------------------------------ ERROR ------------------------------------
 export interface IFieldError {
@@ -17,12 +21,11 @@ export interface IFieldError {
 
 interface IFieldErrors {
   [index: string]: IFieldError;
-};
+}
 
 export interface IError {
   fieldErrors: IFieldErrors;
 }
-
 
 // ------------------------------------ FORM --------------------------------------
 export interface IConstraint {
@@ -31,21 +34,20 @@ export interface IConstraint {
 }
 
 export type IInputChangeHandler = (name: string, value: string, error: IFieldError) => void;
-export type IInputBlurHandler = (name: string, value: string) => void;
-export type IInputFetchHandler = (value: string, onSuccess: (data: any) => void ) => void;
-export type IInputValueHandler = (value: string ) => void;
+export type IInputBlurHandler   = (name: string, value: string) => void;
+export type IInputFetchHandler  = (value: string, onSuccess: (data: any) => void) => void;
+export type IInputValueHandler  = (value: string) => void;
 
 export interface ISelectOption {
-  value: string|number;
+  value: string | number;
   name: string;
-};
+}
 
-// ------------------------------------ MODEL .------------------------------------
-
+// ------------------------------------ MODEL ------------------------------------
 interface IBaseEntity {
   id: number;
   isNew?: boolean;
-};
+}
 
 interface INamedEntity extends IBaseEntity {
   name: string;
@@ -59,10 +61,9 @@ interface IPerson extends IBaseEntity {
 export interface IVisit extends IBaseEntity {
   date: Date;
   description: string;
-};
+}
 
-export interface IPetType extends INamedEntity {
-};
+export interface IPetType extends INamedEntity {}
 
 export type IPetTypeId = number;
 
@@ -70,7 +71,7 @@ export interface IPet extends INamedEntity {
   birthDate: Date;
   type: IPetType;
   visits: IVisit[];
-};
+}
 
 // TODO
 export interface IEditablePet extends INamedEntity {
@@ -98,15 +99,14 @@ export interface IPetRequest {
 export interface IOwner extends IPerson {
   address: string;
   city: string;
-  state?: string,
-  zipCode?: string,
+  state?: string;
+  zipCode?: string;
   telephone: string;
   pets: IPet[];
-};
+}
 
-export interface ISpecialty extends INamedEntity {
-};
+export interface ISpecialty extends INamedEntity {}
 
 export interface IVet extends IPerson {
   specialties: ISpecialty[];
-};
+}
