@@ -1,9 +1,8 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+const router = express.Router();
 
-/* Root endpoint */
-router.get('/', function(req, res, next) {
-  res.json({"service":"react-petclinic-server"});
+router.get('/', function (req, res) {
+  res.json({ service: 'react-petclinic-server' });
 });
 
 module.exports = router;
