@@ -3,19 +3,16 @@ package org.springframework.samples.petclinic.model;
 import java.util.HashSet;
 import java.util.Set;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.springframework.data.domain.Persistable;
 
 @Entity
 @Table(name = "users")
-public class User {
+public class User implements Persistable<String> {
 
     @Id
     @Column(name = "username")
@@ -27,48 +24,55 @@ public class User {
     @Column(name = "enabled")
     private Boolean enabled;
 
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "user", fetch = FetchType.EAGER)
-    private Set<Role> roles;
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "user", fetch = FetchType.EAGER, orphanRemoval = true)
+    @NotNull(message = "may not be null")
+    @NotEmpty(message = "may not be empty")
+    private Set<Role> roles = new HashSet<>();
 
-    public String getUsername() {
-        return username;
+    // ---- Persistable support ----
+    @Transient
+    @JsonIgnore
+    private boolean isNew = true;
+
+    @PostLoad
+    @PostPersist
+    void markNotNew() {
+        this.isNew = false;
     }
 
-    public void setUsername(String username) {
-        this.username = username;
+    @Override
+    @JsonIgnore
+    public String getId() {
+        return this.username;
     }
 
-    public String getPassword() {
-        return password;
+    @Override
+    @JsonIgnore
+    public boolean isNew() {
+        return this.isNew;
     }
+    // -----------------------------
 
-    public void setPassword(String password) {
-        this.password = password;
-    }
+    public String getUsername() { return username; }
+    public void setUsername(String username) { this.username = username; }
 
-    public Boolean getEnabled() {
-        return enabled;
-    }
+    public String getPassword() { return password; }
+    public void setPassword(String password) { this.password = password; }
 
-    public void setEnabled(Boolean enabled) {
-        this.enabled = enabled;
-    }
+    public Boolean getEnabled() { return enabled; }
+    public void setEnabled(Boolean enabled) { this.enabled = enabled; }
 
-    public Set<Role> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(Set<Role> roles) {
-        this.roles = roles;
-    }
+    public Set<Role> getRoles() { return roles; }
+    public void setRoles(Set<Role> roles) { this.roles = roles; }
 
     @JsonIgnore
     public void addRole(String roleName) {
-        if(this.roles == null) {
+        if (this.roles == null) {
             this.roles = new HashSet<>();
         }
         Role role = new Role();
         role.setName(roleName);
+        role.setUser(this); // garantir o lado dono
         this.roles.add(role);
     }
 }
