@@ -13,69 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+
 package org.springframework.samples.petclinic.repository;
 
 import java.util.Collection;
 
 import org.springframework.dao.DataAccessException;
-import org.springframework.samples.petclinic.model.BaseEntity;
 import org.springframework.samples.petclinic.model.Owner;
 
-/**
- * Repository class for <code>Owner</code> domain objects All method names are compliant with Spring Data naming
- * conventions so this interface can easily be extended for Spring Data See here: http://static.springsource.org/spring-data/jpa/docs/current/reference/html/jpa.repositories.html#jpa.query-methods.query-creation
- *
- * @author Ken Krebs
- * @author Juergen Hoeller
- * @author Sam Brannen
- * @author Michael Isvy
- * @author Vitaliy Fedoriv
- */
 public interface OwnerRepository {
 
-    /**
-     * Retrieve <code>Owner</code>s from the data store by last name, returning all owners whose last name <i>starts</i>
-     * with the given name.
-     *
-     * @param lastName Value to search for
-     * @return a <code>Collection</code> of matching <code>Owner</code>s (or an empty <code>Collection</code> if none
-     * found)
-     */
     Collection<Owner> findByLastName(String lastName) throws DataAccessException;
 
-    /**
-     * Retrieve an <code>Owner</code> from the data store by id.
-     *
-     * @param id the id to search for
-     * @return the <code>Owner</code> if found
-     * @throws org.springframework.dao.DataRetrievalFailureException if not found
-     */
     Owner findById(int id) throws DataAccessException;
 
-
-    /**
-     * Save an <code>Owner</code> to the data store, either inserting or updating it.
-     *
-     * @param owner the <code>Owner</code> to save
-     * @see BaseEntity#isNew
-     */
     void save(Owner owner) throws DataAccessException;
-    
-    /**
-     * Retrieve <code>Owner</code>s from the data store, returning all owners 
-     *
-     * @return a <code>Collection</code> of <code>Owner</code>s (or an empty <code>Collection</code> if none
-     * found)
-     */
-	Collection<Owner> findAll() throws DataAccessException;
-	
-    /**
-     * Delete an <code>Owner</code> to the data store by <code>Owner</code>.
-     *
-     * @param owner the <code>Owner</code> to delete
-     * 
-     */
-	void delete(Owner owner) throws DataAccessException;
 
+    Collection<Owner> findAll() throws DataAccessException;
 
+    void delete(Owner owner) throws DataAccessException;
 }
