@@ -1,19 +1,3 @@
-/*
- * Copyright 2016-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.springframework.samples.petclinic.repository.springdatajpa;
 
 import jakarta.persistence.EntityManager;
@@ -21,23 +5,31 @@ import jakarta.persistence.PersistenceContext;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.samples.petclinic.model.Pet;
-
-/**
- * @author Vitaliy Fedoriv
- *
- */
+import org.springframework.transaction.annotation.Transactional;
 
 @Profile("spring-data-jpa")
 public class SpringDataPetRepositoryImpl implements PetRepositoryOverride {
-	
-	@PersistenceContext
+
+    @PersistenceContext
     private EntityManager em;
 
-	@Override
-	public void delete(Pet pet) {
-		String petId = pet.getId().toString();
-		this.em.createQuery("DELETE FROM Visit visit WHERE pet_id=" + petId).executeUpdate();
-		this.em.createQuery("DELETE FROM Pet pet WHERE id=" + petId).executeUpdate();
-	}
+    @Transactional
+    @Override
+    public void delete(Pet pet) {
+        Integer petId = pet.getId();
 
+        // 1) Apaga visitas do pet
+        em.createQuery("DELETE FROM Visit v WHERE v.pet.id = :petId")
+          .setParameter("petId", petId)
+          .executeUpdate();
+
+        // 2) Apaga o próprio pet
+        em.createQuery("DELETE FROM Pet p WHERE p.id = :petId")
+          .setParameter("petId", petId)
+          .executeUpdate();
+
+        // 3) Garante que o 1º nível de cache não devolva entidade obsoleta
+        em.flush();
+        em.clear();
+    }
 }
