@@ -1,59 +1,50 @@
-/*
- * Copyright 2016-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.springframework.samples.petclinic.repository.springdatajpa;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 
 import org.springframework.context.annotation.Profile;
-import org.springframework.samples.petclinic.model.Pet;
 import org.springframework.samples.petclinic.model.PetType;
-import org.springframework.samples.petclinic.model.Visit;
-
-/**
- * @author Vitaliy Fedoriv
- *
- */
+import org.springframework.transaction.annotation.Transactional;
 
 @Profile("spring-data-jpa")
 public class SpringDataPetTypeRepositoryImpl implements PetTypeRepositoryOverride {
-	
-	@PersistenceContext
+
+    @PersistenceContext
     private EntityManager em;
 
-	@SuppressWarnings("unchecked")
-	@Override
-	public void delete(PetType petType) {
-		String petTypeId = petType.getId().toString();
-		
-		List<Pet> pets = new ArrayList<Pet>();
-		pets = this.em.createQuery("SELECT pet FROM Pet pet WHERE type_id=" + petTypeId).getResultList();
-		for (Pet pet : pets){
-			List<Visit> visits = new ArrayList<Visit>();
-			visits = pet.getVisits();
-			for (Visit visit : visits){
-				this.em.createQuery("DELETE FROM Visit visit WHERE id=" + visit.getId().toString()).executeUpdate();
-			}
-			this.em.createQuery("DELETE FROM Pet pet WHERE id=" + pet.getId().toString()).executeUpdate();
-		}
-		this.em.createQuery("DELETE FROM PetType pettype WHERE id=" + petTypeId).executeUpdate();
-	}
+    @Transactional
+    @Override
+    public void delete(PetType petType) {
+        Integer typeId = petType.getId();
 
+        // 1) Descobrir IDs dos pets desse tipo
+        List<Integer> petIds = em.createQuery(
+                "SELECT p.id FROM Pet p WHERE p.type.id = :typeId", Integer.class)
+            .setParameter("typeId", typeId)
+            .getResultList();
+
+        if (!petIds.isEmpty()) {
+            // 2) Apagar visitas desses pets
+            em.createQuery("DELETE FROM Visit v WHERE v.pet.id IN :petIds")
+              .setParameter("petIds", petIds)
+              .executeUpdate();
+
+            // 3) Apagar os pets
+            em.createQuery("DELETE FROM Pet p WHERE p.id IN :petIds")
+              .setParameter("petIds", petIds)
+              .executeUpdate();
+        }
+
+        // 4) Finalmente, apagar o tipo
+        em.createQuery("DELETE FROM PetType pt WHERE pt.id = :typeId")
+          .setParameter("typeId", typeId)
+          .executeUpdate();
+
+        // 5) Sincroniza e limpa o contexto
+        em.flush();
+        em.clear();
+    }
 }
