@@ -1,19 +1,3 @@
-/*
- * Copyright 2016-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.springframework.samples.petclinic.rest;
 
 import java.util.Collection;
@@ -34,110 +18,118 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import co.elastic.apm.api.ElasticApm;
-import co.elastic.apm.api.Transaction;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
 
 /**
- * @author Vitaliy Fedoriv
- *
+ * Owner REST Controller
  */
-
 @RestController
 @CrossOrigin(exposedHeaders = "errors, content-type")
 @RequestMapping("/api/owners")
 public class OwnerRestController {
 
-	@Autowired
-	private ClinicService clinicService;
+    private final ClinicService clinicService;
 
-	@RequestMapping(value = "/*/lastname/{lastName}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Collection<Owner>> getOwnersList(@PathVariable("lastName") String ownerLastName) {
-		if (ownerLastName == null) {
-			ownerLastName = "";
-		}
-		Collection<Owner> owners = this.clinicService.findOwnerByLastName(ownerLastName);
-		if (owners.isEmpty()) {
-			return new ResponseEntity<Collection<Owner>>(HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<Collection<Owner>>(owners, HttpStatus.OK);
-	}
+    public OwnerRestController(ClinicService clinicService) {
+        this.clinicService = clinicService;
+    }
 
-	@RequestMapping(value = "", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Collection<Owner>> getOwners() {
-		Collection<Owner> owners = this.clinicService.findAllOwners();
-		if (owners.isEmpty()) {
-			return new ResponseEntity<Collection<Owner>>(HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<Collection<Owner>>(owners, HttpStatus.OK);
-	}
+    // Busca por sobrenome via path variable: /api/owners/lastname/{lastName}
+    @GetMapping(value = "/lastname/{lastName}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Collection<Owner>> getOwnersList(@PathVariable("lastName") String ownerLastName) {
+        if (ownerLastName == null) {
+            ownerLastName = "";
+        }
+        Collection<Owner> owners = this.clinicService.findOwnerByLastName(ownerLastName);
+        if (owners.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(owners, HttpStatus.OK);
+    }
 
-	@RequestMapping(value = "/{ownerId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Owner> getOwner(@PathVariable("ownerId") int ownerId) {
-		Owner owner = null;
-		owner = this.clinicService.findOwnerById(ownerId);
-		if (owner == null) {
-			return new ResponseEntity<Owner>(HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<Owner>(owner, HttpStatus.OK);
-	}
+    // Lista todos
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Collection<Owner>> getOwners() {
+        Collection<Owner> owners = this.clinicService.findAllOwners();
+        if (owners.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(owners, HttpStatus.OK);
+    }
 
-	@CaptureSpan(value = "addOwner")
-	@RequestMapping(value = "", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Owner> addOwner(@RequestBody @Valid Owner owner, BindingResult bindingResult,
-			UriComponentsBuilder ucBuilder) {
-        Transaction transaction = ElasticApm.currentTransaction();
-		BindingErrorsResponse errors = new BindingErrorsResponse();
-		HttpHeaders headers = new HttpHeaders();
-		if (bindingResult.hasErrors() || (owner == null)) {
-			errors.addAllErrors(bindingResult);
-			headers.add("errors", errors.toJSON());
-			return new ResponseEntity<Owner>(headers, HttpStatus.BAD_REQUEST);
-		}
-		this.clinicService.saveOwner(owner);
-		headers.setLocation(ucBuilder.path("/api/owners/{id}").buildAndExpand(owner.getId()).toUri());
-		return new ResponseEntity<Owner>(owner, headers, HttpStatus.CREATED);
-	}
+    // Busca por id
+    @GetMapping(value = "/{ownerId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Owner> getOwner(@PathVariable("ownerId") int ownerId) {
+        Owner owner = this.clinicService.findOwnerById(ownerId);
+        if (owner == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(owner, HttpStatus.OK);
+    }
 
-    @CaptureSpan(value = "updateOwner")
-	@RequestMapping(value = "/{ownerId}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Owner> updateOwner(@PathVariable("ownerId") int ownerId, @RequestBody @Valid Owner owner,
-			BindingResult bindingResult, UriComponentsBuilder ucBuilder) {
-
+    // Criação
+    @CaptureSpan(value = "addOwner")
+    @PostMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Owner> addOwner(@RequestBody @Valid Owner owner,
+                                          BindingResult bindingResult,
+                                          UriComponentsBuilder ucBuilder) {
         BindingErrorsResponse errors = new BindingErrorsResponse();
-		HttpHeaders headers = new HttpHeaders();
-		if (bindingResult.hasErrors() || (owner == null)) {
-			errors.addAllErrors(bindingResult);
-			headers.add("errors", errors.toJSON());
-			return new ResponseEntity<Owner>(headers, HttpStatus.BAD_REQUEST);
-		}
-		Owner currentOwner = this.clinicService.findOwnerById(ownerId);
-		if (currentOwner == null) {
-			return new ResponseEntity<Owner>(HttpStatus.NOT_FOUND);
-		}
-		currentOwner.setAddress(owner.getAddress());
-		currentOwner.setCity(owner.getCity());
+        HttpHeaders headers = new HttpHeaders();
+        if (bindingResult.hasErrors() || (owner == null)) {
+            errors.addAllErrors(bindingResult);
+            headers.add("errors", errors.toJSON());
+            return new ResponseEntity<>(headers, HttpStatus.BAD_REQUEST);
+        }
+        this.clinicService.saveOwner(owner);
+        headers.setLocation(ucBuilder.path("/api/owners/{id}").buildAndExpand(owner.getId()).toUri());
+        return new ResponseEntity<>(owner, headers, HttpStatus.CREATED);
+    }
+
+    // Atualização
+    @CaptureSpan(value = "updateOwner")
+    @PutMapping(value = "/{ownerId}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Owner> updateOwner(@PathVariable("ownerId") int ownerId,
+                                             @RequestBody @Valid Owner owner,
+                                             BindingResult bindingResult,
+                                             UriComponentsBuilder ucBuilder) {
+        BindingErrorsResponse errors = new BindingErrorsResponse();
+        HttpHeaders headers = new HttpHeaders();
+        if (bindingResult.hasErrors() || (owner == null)) {
+            errors.addAllErrors(bindingResult);
+            headers.add("errors", errors.toJSON());
+            return new ResponseEntity<>(headers, HttpStatus.BAD_REQUEST);
+        }
+        Owner currentOwner = this.clinicService.findOwnerById(ownerId);
+        if (currentOwner == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        currentOwner.setAddress(owner.getAddress());
+        currentOwner.setCity(owner.getCity());
         currentOwner.setZipCode(owner.getZipCode());
         currentOwner.setState(owner.getState());
-		currentOwner.setFirstName(owner.getFirstName());
-		currentOwner.setLastName(owner.getLastName());
-		currentOwner.setTelephone(owner.getTelephone());
-		this.clinicService.saveOwner(currentOwner);
-		return new ResponseEntity<Owner>(currentOwner, HttpStatus.NO_CONTENT);
-	}
+        currentOwner.setFirstName(owner.getFirstName());
+        currentOwner.setLastName(owner.getLastName());
+        currentOwner.setTelephone(owner.getTelephone());
+        this.clinicService.saveOwner(currentOwner);
+        // Mantemos 204 No Content como no seu contrato de testes
+        return new ResponseEntity<>(currentOwner, HttpStatus.NO_CONTENT);
+    }
 
-	@RequestMapping(value = "/{ownerId}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	@Transactional
-	public ResponseEntity<Void> deleteOwner(@PathVariable("ownerId") int ownerId) {
-		Owner owner = this.clinicService.findOwnerById(ownerId);
-		if (owner == null) {
-			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
-		}
-		this.clinicService.deleteOwner(owner);
-		return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
-	}
-
+    // Exclusão
+    @DeleteMapping(value = "/{ownerId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    @Transactional
+    public ResponseEntity<Void> deleteOwner(@PathVariable("ownerId") int ownerId) {
+        Owner owner = this.clinicService.findOwnerById(ownerId);
+        if (owner == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        this.clinicService.deleteOwner(owner);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
 }
