@@ -13,6 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+
 package org.springframework.samples.petclinic.rest;
 
 import java.util.ArrayList;
@@ -34,96 +36,106 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 /**
  * @author Vitaliy Fedoriv
- *
  */
 
 @RestController
-@CrossOrigin(exposedHeaders = "errors, content-type")
-@RequestMapping("api/vets")
+@CrossOrigin(exposedHeaders = { "errors", "content-type" })
+@RequestMapping("/api/vets")
 public class VetRestController {
 
-	@Autowired
-	private ClinicService clinicService;
+    @Autowired
+    private ClinicService clinicService;
 
-    @PreAuthorize( "hasRole(@roles.VET_ADMIN)" )
-	@RequestMapping(value = "", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Collection<Vet>> getAllVets(){
-		Collection<Vet> vets = new ArrayList<Vet>();
-		vets.addAll(this.clinicService.findAllVets());
-		if (vets.isEmpty()){
-			return new ResponseEntity<Collection<Vet>>(HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<Collection<Vet>>(vets, HttpStatus.OK);
-	}
+    @PreAuthorize("hasRole('VET_ADMIN')")
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Collection<Vet>> getAllVets() {
+        Collection<Vet> vets = new ArrayList<>();
+        vets.addAll(this.clinicService.findAllVets());
+        if (vets.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(vets, HttpStatus.OK);
+    }
 
-    @PreAuthorize( "hasRole(@roles.VET_ADMIN)" )
-	@RequestMapping(value = "/{vetId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Vet> getVet(@PathVariable("vetId") int vetId){
-		Vet vet = this.clinicService.findVetById(vetId);
-		if(vet == null){
-			return new ResponseEntity<Vet>(HttpStatus.NOT_FOUND);
-		}
-		return new ResponseEntity<Vet>(vet, HttpStatus.OK);
-	}
+    @PreAuthorize("hasRole('VET_ADMIN')")
+    @GetMapping(value = "/{vetId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Vet> getVet(@PathVariable("vetId") int vetId) {
+        Vet vet = this.clinicService.findVetById(vetId);
+        if (vet == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(vet, HttpStatus.OK);
+    }
 
-    @PreAuthorize( "hasRole(@roles.VET_ADMIN)" )
-	@RequestMapping(value = "", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Vet> addVet(@RequestBody @Valid Vet vet, BindingResult bindingResult, UriComponentsBuilder ucBuilder){
-		BindingErrorsResponse errors = new BindingErrorsResponse();
-		HttpHeaders headers = new HttpHeaders();
-		if(bindingResult.hasErrors() || (vet == null)){
-			errors.addAllErrors(bindingResult);
-			headers.add("errors", errors.toJSON());
-			return new ResponseEntity<Vet>(headers, HttpStatus.BAD_REQUEST);
-		}
-		this.clinicService.saveVet(vet);
-		headers.setLocation(ucBuilder.path("/api/vets/{id}").buildAndExpand(vet.getId()).toUri());
-		return new ResponseEntity<Vet>(vet, headers, HttpStatus.CREATED);
-	}
+    @PreAuthorize("hasRole('VET_ADMIN')")
+    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Vet> addVet(@RequestBody @Valid Vet vet,
+                                      BindingResult bindingResult,
+                                      UriComponentsBuilder ucBuilder) {
+        BindingErrorsResponse errors = new BindingErrorsResponse();
+        HttpHeaders headers = new HttpHeaders();
+        if (bindingResult.hasErrors() || vet == null) {
+            errors.addAllErrors(bindingResult);
+            headers.add("errors", errors.toJSON());
+            return new ResponseEntity<>(headers, HttpStatus.BAD_REQUEST);
+        }
+        this.clinicService.saveVet(vet);
+        headers.setLocation(ucBuilder.path("/api/vets/{id}").buildAndExpand(vet.getId()).toUri());
+        // Alternativa idiomática:
+        // return ResponseEntity.created(headers.getLocation()).body(vet);
+        return new ResponseEntity<>(vet, headers, HttpStatus.CREATED);
+    }
 
-    @PreAuthorize( "hasRole(@roles.VET_ADMIN)" )
-	@RequestMapping(value = "/{vetId}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	public ResponseEntity<Vet> updateVet(@PathVariable("vetId") int vetId, @RequestBody @Valid Vet vet, BindingResult bindingResult){
-		BindingErrorsResponse errors = new BindingErrorsResponse();
-		HttpHeaders headers = new HttpHeaders();
-		if(bindingResult.hasErrors() || (vet == null)){
-			errors.addAllErrors(bindingResult);
-			headers.add("errors", errors.toJSON());
-			return new ResponseEntity<Vet>(headers, HttpStatus.BAD_REQUEST);
-		}
-		Vet currentVet = this.clinicService.findVetById(vetId);
-		if(currentVet == null){
-			return new ResponseEntity<Vet>(HttpStatus.NOT_FOUND);
-		}
-		currentVet.setFirstName(vet.getFirstName());
-		currentVet.setLastName(vet.getLastName());
-		currentVet.clearSpecialties();
-		for(Specialty spec : vet.getSpecialties()) {
-			currentVet.addSpecialty(spec);
-		}
-		this.clinicService.saveVet(currentVet);
-		return new ResponseEntity<Vet>(currentVet, HttpStatus.NO_CONTENT);
-	}
+    @PreAuthorize("hasRole('VET_ADMIN')")
+    @PutMapping(value = "/{vetId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Vet> updateVet(@PathVariable("vetId") int vetId,
+                                         @RequestBody @Valid Vet vet,
+                                         BindingResult bindingResult) {
+        BindingErrorsResponse errors = new BindingErrorsResponse();
+        HttpHeaders headers = new HttpHeaders();
+        if (bindingResult.hasErrors() || vet == null) {
+            errors.addAllErrors(bindingResult);
+            headers.add("errors", errors.toJSON());
+            return new ResponseEntity<>(headers, HttpStatus.BAD_REQUEST);
+        }
 
-    @PreAuthorize( "hasRole(@roles.VET_ADMIN)" )
-	@RequestMapping(value = "/{vetId}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-	@Transactional
-	public ResponseEntity<Void> deleteVet(@PathVariable("vetId") int vetId){
-		Vet vet = this.clinicService.findVetById(vetId);
-		if(vet == null){
-			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
-		}
-		this.clinicService.deleteVet(vet);
-		return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
-	}
+        Vet currentVet = this.clinicService.findVetById(vetId);
+        if (currentVet == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
 
+        currentVet.setFirstName(vet.getFirstName());
+        currentVet.setLastName(vet.getLastName());
+        currentVet.clearSpecialties();
+        for (Specialty spec : vet.getSpecialties()) {
+            currentVet.addSpecialty(spec);
+        }
+        this.clinicService.saveVet(currentVet);
 
+        // 204 No Content não deve retornar body. Se quiser retornar o recurso atualizado, use 200 OK.
+        // return ResponseEntity.noContent().build();
+        return new ResponseEntity<>(currentVet, HttpStatus.OK);
+    }
 
+    @PreAuthorize("hasRole('VET_ADMIN')")
+    @DeleteMapping(value = "/{vetId}")
+    @Transactional
+    public ResponseEntity<Void> deleteVet(@PathVariable("vetId") int vetId) {
+        Vet vet = this.clinicService.findVetById(vetId);
+        if (vet == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        this.clinicService.deleteVet(vet);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
 }
