@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright 2002-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,30 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- */
-package org.springframework.samples.petclinic.service.clinicService;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Collection;
-import java.util.Date;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.MockitoAnnotations;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.samples.petclinic.model.Owner;
-import org.springframework.samples.petclinic.model.Pet;
-import org.springframework.samples.petclinic.model.PetType;
-import org.springframework.samples.petclinic.model.Specialty;
-import org.springframework.samples.petclinic.model.Vet;
-import org.springframework.samples.petclinic.model.Visit;
-import org.springframework.samples.petclinic.service.ClinicService;
-import org.springframework.samples.petclinic.util.EntityUtils;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
+ *
  * <p> Base class for {@link ClinicService} integration tests. </p> <p> Subclasses should specify Spring context
  * configuration using {@link ContextConfiguration @ContextConfiguration} annotation </p> <p>
  * AbstractclinicServiceTests and its subclasses benefit from the following services provided by the Spring
@@ -55,14 +32,36 @@ import org.springframework.transaction.annotation.Transactional;
  * @author Michael Isvy
  * @author Vitaliy Fedoriv
  */
+
+package org.springframework.samples.petclinic.service.clinicService;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Collection;
+import java.util.Date;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.samples.petclinic.model.Owner;
+import org.springframework.samples.petclinic.model.Pet;
+import org.springframework.samples.petclinic.model.PetType;
+import org.springframework.samples.petclinic.model.Specialty;
+import org.springframework.samples.petclinic.model.Vet;
+import org.springframework.samples.petclinic.model.Visit;
+import org.springframework.samples.petclinic.service.ClinicService;
+import org.springframework.samples.petclinic.util.EntityUtils;
+import org.springframework.transaction.annotation.Transactional;
+
 public abstract class AbstractClinicServiceTests {
 
     @Autowired
     protected ClinicService clinicService;
 
-    @Before
+    @BeforeEach
     public void init() {
-        MockitoAnnotations.initMocks(this);
+        // Nenhuma inicialização necessária: usamos o ClinicService real do contexto de teste.
+        // Removido MockitoAnnotations.initMocks(this) (obsoleto e desnecessário aqui).
     }
 
     @Test
@@ -124,18 +123,7 @@ public abstract class AbstractClinicServiceTests {
         Pet pet7 = this.clinicService.findPetById(7);
         assertThat(pet7.getName()).startsWith("Samantha");
         assertThat(pet7.getOwner().getFirstName()).isEqualTo("Jean");
-
     }
-
-//    @Test
-//    public void shouldFindAllPetTypes() {
-//        Collection<PetType> petTypes = this.clinicService.findPetTypes();
-//
-//        PetType petType1 = EntityUtils.getById(petTypes, PetType.class, 1);
-//        assertThat(petType1.getName()).isEqualTo("cat");
-//        PetType petType4 = EntityUtils.getById(petTypes, PetType.class, 4);
-//        assertThat(petType4.getName()).isEqualTo("snake");
-//    }
 
     @Test
     @Transactional
@@ -162,7 +150,7 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldUpdatePetName() throws Exception {
+    public void shouldUpdatePetName() {
         Pet pet7 = this.clinicService.findPetById(7);
         String oldName = pet7.getName();
 
@@ -202,7 +190,7 @@ public abstract class AbstractClinicServiceTests {
     }
 
     @Test
-       public void shouldFindVisitsByPetId() throws Exception {
+    public void shouldFindVisitsByPetId() {
         Collection<Visit> visits = this.clinicService.findVisitsByPetId(7);
         assertThat(visits.size()).isEqualTo(2);
         Visit[] visitArr = visits.toArray(new Visit[visits.size()]);
@@ -212,7 +200,7 @@ public abstract class AbstractClinicServiceTests {
     }
 
     @Test
-    public void shouldFindAllPets(){
+    public void shouldFindAllPets() {
         Collection<Pet> pets = this.clinicService.findAllPets();
         Pet pet1 = EntityUtils.getById(pets, Pet.class, 1);
         assertThat(pet1.getName()).isEqualTo("Leo");
@@ -222,26 +210,26 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeletePet(){
+    public void shouldDeletePet() {
         Pet pet = this.clinicService.findPetById(1);
         this.clinicService.deletePet(pet);
         try {
-        pet = this.clinicService.findPetById(1);
-		} catch (Exception e) {
-			pet = null;
-		}
+            pet = this.clinicService.findPetById(1);
+        } catch (Exception e) {
+            pet = null;
+        }
         assertThat(pet).isNull();
     }
 
     @Test
-    public void shouldFindVisitDyId(){
-    	Visit visit = this.clinicService.findVisitById(1);
-    	assertThat(visit.getId()).isEqualTo(1);
-    	assertThat(visit.getPet().getName()).isEqualTo("Samantha");
+    public void shouldFindVisitById() {
+        Visit visit = this.clinicService.findVisitById(1);
+        assertThat(visit.getId()).isEqualTo(1);
+        assertThat(visit.getPet().getName()).isEqualTo("Samantha");
     }
 
     @Test
-    public void shouldFindAllVisits(){
+    public void shouldFindAllVisits() {
         Collection<Visit> visits = this.clinicService.findAllVisits();
         Visit visit1 = EntityUtils.getById(visits, Visit.class, 1);
         assertThat(visit1.getPet().getName()).isEqualTo("Samantha");
@@ -262,7 +250,6 @@ public abstract class AbstractClinicServiceTests {
         visit.setDate(new Date());
         visit.setDescription("new visit");
 
-
         this.clinicService.saveVisit(visit);
         assertThat(visit.getId().longValue()).isNotEqualTo(0);
 
@@ -272,9 +259,9 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldUpdateVisit(){
-    	Visit visit = this.clinicService.findVisitById(1);
-    	String oldDesc = visit.getDescription();
+    public void shouldUpdateVisit() {
+        Visit visit = this.clinicService.findVisitById(1);
+        String oldDesc = visit.getDescription();
         String newDesc = oldDesc + "X";
         visit.setDescription(newDesc);
         this.clinicService.saveVisit(visit);
@@ -284,22 +271,22 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeleteVisit(){
-    	Visit visit = this.clinicService.findVisitById(1);
+    public void shouldDeleteVisit() {
+        Visit visit = this.clinicService.findVisitById(1);
         this.clinicService.deleteVisit(visit);
         try {
-        	visit = this.clinicService.findVisitById(1);
-		} catch (Exception e) {
-			visit = null;
-		}
+            visit = this.clinicService.findVisitById(1);
+        } catch (Exception e) {
+            visit = null;
+        }
         assertThat(visit).isNull();
     }
 
     @Test
-    public void shouldFindVetDyId(){
-    	Vet vet = this.clinicService.findVetById(1);
-    	assertThat(vet.getFirstName()).isEqualTo("James");
-    	assertThat(vet.getLastName()).isEqualTo("Carter");
+    public void shouldFindVetById() {
+        Vet vet = this.clinicService.findVetById(1);
+        assertThat(vet.getFirstName()).isEqualTo("James");
+        assertThat(vet.getLastName()).isEqualTo("Carter");
     }
 
     @Test
@@ -321,9 +308,9 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldUpdateVet(){
-    	Vet vet = this.clinicService.findVetById(1);
-    	String oldLastName = vet.getLastName();
+    public void shouldUpdateVet() {
+        Vet vet = this.clinicService.findVetById(1);
+        String oldLastName = vet.getLastName();
         String newLastName = oldLastName + "X";
         vet.setLastName(newLastName);
         this.clinicService.saveVet(vet);
@@ -333,19 +320,19 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeleteVet(){
-    	Vet vet = this.clinicService.findVetById(1);
+    public void shouldDeleteVet() {
+        Vet vet = this.clinicService.findVetById(1);
         this.clinicService.deleteVet(vet);
         try {
-        	vet = this.clinicService.findVetById(1);
-		} catch (Exception e) {
-			vet = null;
-		}
+            vet = this.clinicService.findVetById(1);
+        } catch (Exception e) {
+            vet = null;
+        }
         assertThat(vet).isNull();
     }
 
     @Test
-    public void shouldFindAllOwners(){
+    public void shouldFindAllOwners() {
         Collection<Owner> owners = this.clinicService.findAllOwners();
         Owner owner1 = EntityUtils.getById(owners, Owner.class, 1);
         assertThat(owner1.getFirstName()).isEqualTo("George");
@@ -355,25 +342,25 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeleteOwner(){
-    	Owner owner = this.clinicService.findOwnerById(1);
+    public void shouldDeleteOwner() {
+        Owner owner = this.clinicService.findOwnerById(1);
         this.clinicService.deleteOwner(owner);
         try {
-        	owner = this.clinicService.findOwnerById(1);
-		} catch (Exception e) {
-			owner = null;
-		}
+            owner = this.clinicService.findOwnerById(1);
+        } catch (Exception e) {
+            owner = null;
+        }
         assertThat(owner).isNull();
     }
 
     @Test
-    public void shouldFindPetTypeById(){
-    	PetType petType = this.clinicService.findPetTypeById(1);
-    	assertThat(petType.getName()).isEqualTo("cat");
+    public void shouldFindPetTypeById() {
+        PetType petType = this.clinicService.findPetTypeById(1);
+        assertThat(petType.getName()).isEqualTo("cat");
     }
 
     @Test
-    public void shouldFindAllPetTypes(){
+    public void shouldFindAllPetTypes() {
         Collection<PetType> petTypes = this.clinicService.findAllPetTypes();
         PetType petType1 = EntityUtils.getById(petTypes, PetType.class, 1);
         assertThat(petType1.getName()).isEqualTo("cat");
@@ -399,9 +386,9 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldUpdatePetType(){
-    	PetType petType = this.clinicService.findPetTypeById(1);
-    	String oldLastName = petType.getName();
+    public void shouldUpdatePetType() {
+        PetType petType = this.clinicService.findPetTypeById(1);
+        String oldLastName = petType.getName();
         String newLastName = oldLastName + "X";
         petType.setName(newLastName);
         this.clinicService.savePetType(petType);
@@ -411,25 +398,25 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeletePetType(){
-    	PetType petType = this.clinicService.findPetTypeById(1);
+    public void shouldDeletePetType() {
+        PetType petType = this.clinicService.findPetTypeById(1);
         this.clinicService.deletePetType(petType);
         try {
-        	petType = this.clinicService.findPetTypeById(1);
-		} catch (Exception e) {
-			petType = null;
-		}
+            petType = this.clinicService.findPetTypeById(1);
+        } catch (Exception e) {
+            petType = null;
+        }
         assertThat(petType).isNull();
     }
 
     @Test
-    public void shouldFindSpecialtyById(){
-    	Specialty specialty = this.clinicService.findSpecialtyById(1);
-    	assertThat(specialty.getName()).isEqualTo("radiology");
+    public void shouldFindSpecialtyById() {
+        Specialty specialty = this.clinicService.findSpecialtyById(1);
+        assertThat(specialty.getName()).isEqualTo("radiology");
     }
 
     @Test
-    public void shouldFindAllSpecialtys(){
+    public void shouldFindAllSpecialtys() {
         Collection<Specialty> specialties = this.clinicService.findAllSpecialties();
         Specialty specialty1 = EntityUtils.getById(specialties, Specialty.class, 1);
         assertThat(specialty1.getName()).isEqualTo("radiology");
@@ -455,9 +442,9 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldUpdateSpecialty(){
-    	Specialty specialty = this.clinicService.findSpecialtyById(1);
-    	String oldLastName = specialty.getName();
+    public void shouldUpdateSpecialty() {
+        Specialty specialty = this.clinicService.findSpecialtyById(1);
+        String oldLastName = specialty.getName();
         String newLastName = oldLastName + "X";
         specialty.setName(newLastName);
         this.clinicService.saveSpecialty(specialty);
@@ -467,17 +454,14 @@ public abstract class AbstractClinicServiceTests {
 
     @Test
     @Transactional
-    public void shouldDeleteSpecialty(){
-    	Specialty specialty = this.clinicService.findSpecialtyById(1);
+    public void shouldDeleteSpecialty() {
+        Specialty specialty = this.clinicService.findSpecialtyById(1);
         this.clinicService.deleteSpecialty(specialty);
         try {
-        	specialty = this.clinicService.findSpecialtyById(1);
-		} catch (Exception e) {
-			specialty = null;
-		}
+            specialty = this.clinicService.findSpecialtyById(1);
+        } catch (Exception e) {
+            specialty = null;
+        }
         assertThat(specialty).isNull();
     }
-
-
-
 }
